feat(infra): durable b-roll queue, per-object media authz, cost precision, export UI
Ships the four items deferred from the review-swarm batch:

- B-roll generation is now a durable queue job ('video_generate'): routed
  through pg-boss when QUEUE_DRIVER=pgboss (retry/backoff/DLQ, 45-min expiry),
  bounded inline otherwise. Jobs interrupted by a restart are recovered on
  boot — resumed by external_task_id (never re-submitted/double-billed) or
  failed cleanly when the provider task id was never recorded.
- Per-object media authorization (fiji's checkVideoAccess pattern): /hls-public,
  /hls-proxy, /video-raw, /video-proxy and /local-storage media keys now allow a
  scoped HMAC path token (minted by the storage adapters; HLS relative segment
  URLs inherit it), public/unlisted projects, or an authenticated
  owner/collaborator — private media is no longer a bare capability URL. Also
  closes the /local-storage any-logged-in-user-reads-any-private-key gap.
  (Supabase public-bucket serving in prod remains bucket-level by design.)
- Migration 046 (applied + db:check healthy): token_usage.cost_cents integer →
  double precision so sub-cent utility calls stop reading as free (aggregations
  now sum as float8), plus composite index (user_id, occurred_at) for the
  rolling-24h generation-cap count.
- Audio Studio Export is now functional end-to-end: keeps render_id, polls the
  render-status endpoint with progress/success/error banners and a real
  download link; Versions drawer download icons link the actual renders.

All 4 packages typecheck clean; backend suite 509/509 (adds mediaToken tests).

+++ podcast-saas/backend-api/src/controllers/admin/v1/users.controller.ts
@@ -66,7 +66,7 @@ export async function registerAdminUsersRoutes(app: FastifyInstance): Promise<vo
       const agg = {
         input:  sql<number>`coalesce(sum(input_tokens),0)::int`,
         output: sql<number>`coalesce(sum(output_tokens),0)::int`,
-        cost:   sql<number>`coalesce(sum(cost_cents),0)::int`,
+        cost:   sql<number>`coalesce(sum(cost_cents),0)::float8`, // fractional cents (migration 046)
       };
       const [totals, providerRows, modelRows, taskRows] = await Promise.all([
         db.select(agg).from(token_usage).where(dateRange),
